+++ core/config.h
@@ -10,8 +10,8 @@ namespace Fresa::Config
 {
     inline const str name = "Proxecto Fresa";
     inline const ui8 version[3] = {0, 3, 6};
-    inline const Vec2<> window_size = Vec2(1024, 720);
-    inline const Vec2<> resolution = Vec2(256, 180);
+    inline const Vec2<ui32> window_size = {1024, 720};
+    inline const Vec2<ui32> resolution = {256, 180};
     
     inline const float timestep = 10.0f;
     inline float game_speed = 1.0f;